remove stack extraction methods from the simulation ethereum client
